+++ js_modules/dagster-ui/packages/ui-core/src/gantt/GanttChartSelectionInput.tsx
@@ -5,8 +5,9 @@ import {RunGraphQueryItem} from './toGraphQueryItems';
 import {NO_STATE} from '../run-selection/AntlrRunSelectionVisitor';
 import {RunSelectionLexer} from '../run-selection/generated/RunSelectionLexer';
 import {RunSelectionParser} from '../run-selection/generated/RunSelectionParser';
-import {InputDiv, SelectionAutoCompleteInput} from '../selection/SelectionAutoCompleteInput';
+import {InputDiv, SelectionAutoCompleteInput} from '../selection/SelectionInput';
 import {createSelectionLinter} from '../selection/createSelectionLinter';
+import {useSelectionInputAutoComplete} from '../selection/useSelectionInputAutoComplete';
 import {weakMapMemoize} from '../util/weakMapMemoize';
 
 export const GanttChartSelectionInput = ({
@@ -18,29 +19,13 @@ export const GanttChartSelectionInput = ({
   value: string;
   onChange: (value: string) => void;
 }) => {
-  const attributesMap = useMemo(() => {
-    const statuses = new Set<string>();
-    const names = new Set<string>();
-
-    items.forEach((item) => {
-      if (item.metadata?.state) {
-        statuses.add(item.metadata.state);
-      } else {
-        statuses.add(NO_STATE);
-      }
-      names.add(item.name);
-    });
-    return {name: Array.from(names), status: Array.from(statuses)};
-  }, [items]);
-
+  const useAutoComplete = useMemo(() => createAutoComplete(items), [items]);
   return (
     <Wrapper>
       <SelectionAutoCompleteInput
         id="run-gantt-chart"
-        nameBase="name"
-        attributesMap={attributesMap}
+        useAutoComplete={useAutoComplete}
         placeholder="Type a step subset"
-        functions={FUNCTIONS}
         linter={getLinter()}
         value={value}
         onChange={onChange}
@@ -48,6 +33,36 @@ export const GanttChartSelectionInput = ({
     </Wrapper>
   );
 };
+
+function createAutoComplete(items: RunGraphQueryItem[]) {
+  return function useAutoComplete(value: string, cursor: number) {
+    const attributesMap = useMemo(() => {
+      const statuses = new Set<string>();
+      const names = new Set<string>();
+
+      items.forEach((item) => {
+        if (item.metadata?.state) {
+          statuses.add(item.metadata.state);
+        } else {
+          statuses.add(NO_STATE);
+        }
+        names.add(item.name);
+      });
+      return {name: Array.from(names), status: Array.from(statuses)};
+    }, []);
+
+    return {
+      autoCompleteResults: useSelectionInputAutoComplete({
+        nameBase: 'name',
+        attributesMap,
+        functions: FUNCTIONS,
+        value,
+        cursor,
+      }),
+      loading: false,
+    };
+  };
+}
 
 const getLinter = weakMapMemoize(() =>
   createSelectionLinter({Lexer: RunSelectionLexer, Parser: RunSelectionParser}),
