+++ js_modules/dagster-ui/packages/ui-core/src/pipelines/OpGraphSelectionInput.tsx
@@ -4,8 +4,9 @@ import styled from 'styled-components';
 import {GraphQueryItem} from '../app/GraphQueryImpl';
 import {OpSelectionLexer} from '../op-selection/generated/OpSelectionLexer';
 import {OpSelectionParser} from '../op-selection/generated/OpSelectionParser';
-import {InputDiv, SelectionAutoCompleteInput} from '../selection/SelectionAutoCompleteInput';
+import {InputDiv, SelectionAutoCompleteInput} from '../selection/SelectionInput';
 import {createSelectionLinter} from '../selection/createSelectionLinter';
+import {useSelectionInputAutoComplete} from '../selection/useSelectionInputAutoComplete';
 import {weakMapMemoize} from '../util/weakMapMemoize';
 
 export const OpGraphSelectionInput = ({
@@ -17,22 +18,13 @@ export const OpGraphSelectionInput = ({
   value: string;
   onChange: (value: string) => void;
 }) => {
-  const attributesMap = useMemo(() => {
-    const names = new Set<string>();
-    items.forEach((item) => {
-      names.add(item.name);
-    });
-    return {name: Array.from(names)};
-  }, [items]);
-
+  const useAutoComplete = useMemo(() => createAutoComplete(items), [items]);
   return (
     <Wrapper>
       <SelectionAutoCompleteInput
         id="op-graph"
-        nameBase="name"
-        attributesMap={attributesMap}
+        useAutoComplete={useAutoComplete}
         placeholder="Type an op subset"
-        functions={FUNCTIONS}
         linter={getLinter()}
         value={value}
         onChange={onChange}
@@ -40,6 +32,29 @@ export const OpGraphSelectionInput = ({
     </Wrapper>
   );
 };
+
+function createAutoComplete(items: GraphQueryItem[]) {
+  return function useAutoComplete(value: string, cursor: number) {
+    const attributesMap = useMemo(() => {
+      const names = new Set<string>();
+      items.forEach((item) => {
+        names.add(item.name);
+      });
+      return {name: Array.from(names)};
+    }, []);
+
+    return {
+      autoCompleteResults: useSelectionInputAutoComplete({
+        nameBase: 'name',
+        attributesMap,
+        functions: FUNCTIONS,
+        value,
+        cursor,
+      }),
+      loading: false,
+    };
+  };
+}
 
 const getLinter = weakMapMemoize(() =>
   createSelectionLinter({Lexer: OpSelectionLexer, Parser: OpSelectionParser}),
